Move interface of table to commons project from SlideFlood. Change name of table interface.
